reorganise non-dataset layers
