+++ ocf_data_sampler/common/lightarray.py
