--- ocf_data_sampler/load/nwp/__init__.py
@@ -1,1 +1,0 @@
-from .nwp import open_nwp
