+++ ocf_data_sampler/load/satellite.py
@@ -4,12 +4,12 @@ import json
 import numpy as np
 import xarray as xr
 
-from ocf_data_sampler.load.open_xarray_tensorstore import open_zarr, open_zarrs
+from ocf_data_sampler.common.tensorstore import open_zarr, open_zarrs
 from ocf_data_sampler.load.utils import (
-    assert_values_unique_increasing,
     get_xr_data_array_from_xr_dataset,
     make_spatial_coords_increasing,
 )
+from ocf_data_sampler.common.indexing import assert_values_unique_increasing
 
 
 def open_sat_data(zarr_path: str | list[str]) -> xr.DataArray:
