+++ ocf_data_sampler/numpy_sample/__init__.py
@@ -1,6 +1,6 @@
 """Conversion from Xarray to NumpySample"""
 
-from .convert import convert_to_numpy_sample
-from .datetime_features import encode_datetimes, get_t0_embedding
+from ..torch_datasets.utils.convert import convert_to_numpy_sample
+from ..features.time_encodings import encode_datetimes, get_t0_embedding
 from .common_types import NumpySample, NumpyBatch, TensorBatch
-from .sun_position import make_sun_position_numpy_sample
+from ..features.solar import make_sun_position_numpy_sample
