+++ ocf_data_sampler/select/__init__.py
@@ -4,4 +4,4 @@ from .find_contiguous_time_periods import (
     intersection_of_multiple_dataframes_of_periods,
 )
 from .location import Location
-from .get_indices_in_sorted_unique import get_indices_in_sorted_unique
+from ..common.indexing import get_indices_in_sorted_unique
