+++ ocf_data_sampler/select/spatial_slice.py
