+++ ocf_data_sampler/torch_datasets/utils/spatial_slice_for_dataset.py
@@ -4,7 +4,7 @@
 from ocf_data_sampler.config import Configuration
 from ocf_data_sampler.select import get_indices_in_sorted_unique
 from ocf_data_sampler.select.location import Location
-from ocf_data_sampler.select.select_spatial_slice import (
+from ocf_data_sampler.select.spatial_slice import (
     select_spatial_slice_pixels,
     select_spatial_slice_pixels_multiple,
 )
