+++ tests/select/test_dropout.py
@@ -3,7 +3,7 @@ import pytest
 import xarray as xr
 
 from ocf_data_sampler.select.dropout import apply_history_dropout
-from ocf_data_sampler.time_utils import minutes
+from ocf_data_sampler.common.time_utils import minutes
 
 
 def test_apply_history_dropout_multiple_timedeltas(da_sample):
