+++ tests/test_lightarray.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from ocf_data_sampler.lightarray import LightDataArray
+from ocf_data_sampler.common.lightarray import LightDataArray
 from ocf_data_sampler.load.generation import open_generation
 from ocf_data_sampler.load.nwp import open_nwp
 from ocf_data_sampler.load.satellite import open_sat_data
